Give live object turns a complete orchestration owner
